Move classes into viewmodel package
