+++ auth/src/main/java/com/firebase/ui/auth/util/ui/FlowHolder.java
@@ -4,6 +4,7 @@ import android.app.Application;
 import android.support.annotation.RestrictTo;
 
 import com.firebase.ui.auth.data.model.FlowParameters;
+import com.firebase.ui.auth.viewmodel.ViewModelBase;
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 public class FlowHolder extends ViewModelBase<FlowParameters> {
