+++ auth/src/main/java/com/firebase/ui/auth/viewmodel/email/RecoverPasswordHandler.java
@@ -7,7 +7,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.RestrictTo;
 
 import com.firebase.ui.auth.data.model.Resource;
-import com.firebase.ui.auth.util.data.AuthViewModelBase;
+import com.firebase.ui.auth.viewmodel.AuthViewModelBase;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
